Moving Indexer to utils submodule
