+++ apis/python/src/tiledbsoma/utils/__init__.py
@@ -1,0 +1,5 @@
+from . import IntIndexer
+
+__all__ = [
+    "IntIndexer",
+]
